+++ packages/react/src/components/notification/Notification.tsx
@@ -8,7 +8,7 @@ import classNames from '../../utils/classNames';
 import { IconInfoCircleFill, IconErrorFill, IconAlertCircleFill, IconCheckCircleFill, IconCross } from '../../icons';
 import { AllElementPropsWithoutRef } from '../../utils/elementTypings';
 import { getPlainTextContent } from '../../utils/getPlainTextContent';
-import { Link, LinkSize } from '../link/Link';
+import { Link } from '../link/Link';
 
 export type NotificationType = 'info' | 'error' | 'alert' | 'success';
 
@@ -68,14 +68,9 @@ type CommonProps = React.PropsWithChildren<
      */
     label?: string | React.ReactNode;
     /**
-     * Link to be displayed in the notification. If not provided, the link will not be displayed.
+     * Link element to be displayed in the notification. If not provided, the link will not be displayed.
      */
-    link?: string;
-    /**
-     * Link text to display. If not provided, the link will be displayed as plain text.
-     * Note: This prop is only used when the `link` prop is provided.
-     */
-    linkText?: string;
+    link?: typeof Link | React.ReactNode;
     /**
      * The aria-label of the notification region
      * @default "Notification"
@@ -216,7 +211,6 @@ export const Notification = React.forwardRef<HTMLDivElement, NotificationProps>(
       type = 'info',
       headingLevel = 2,
       link = undefined,
-      linkText = undefined,
       ...rest
     }: NotificationProps,
     ref,
@@ -318,11 +312,7 @@ export const Notification = React.forwardRef<HTMLDivElement, NotificationProps>(
               </div>
             )}
             {children && <div className={styles.body}>{children}</div>}
-            {link && (
-              <Link href={link as string} size={LinkSize.Medium}>
-                {linkText || getPlainTextContent(link)}
-              </Link>
-            )}
+            {link}
           </div>
           {dismissible && (
             <button