+++ CHANGELOG.md
@@ -24,7 +24,7 @@ Changes that are not related to specific components
 - [Component] What has been changed
 - [Notification] body text size set to --fontsize-body-m all variations
 - [Notification] Heading size and gaps in default/basic size inline updated
-- [Notification] `link` and `linkText` props to enable native links in Notification.
+- [Notification] `link` prop added to enable native link in Notification.
 
 #### Fixed
 
@@ -39,7 +39,7 @@ Changes that are not related to specific components
 #### Added
 
 - [Component] What is added?
-- [Notification] New 
+- [Notification] New example on how to have a link inside notification.
 
 #### Changed